Safari-611.1.10/LayoutTests/fast/canvas/canvas-stroke-empty-fill-expected.txt

Series of tests to ensure correct behaviour for stroke() on an empty fillStyle() (alpha=0)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[4] is 0
PASS imgdata[5] is 128
PASS imgdata[6] is 0
PASS successfullyParsed is true

TEST COMPLETE
